requirements.txt
black==24.2.0
click==8.1.7
dnspython==2.5.0
exceptiongroup==1.2.0
iniconfig==2.0.0
isort==5.13.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pymongo==4.6.1
pymongo_inmemory==0.4.0
pytest==8.0.0
python-dotenv==1.0.1
sentinels==1.0.0
tomli==2.0.1
typing_extensions==4.9.0
